NMEA/Info: replace "double" with "fixed"
Fix the type of LoggerPreTakeoffBuffer::HDOP - it was "bool", which is
obviously wrong.

+++ src/GlideComputerBlackboard.hpp
@@ -78,7 +78,7 @@ protected:
   void Initialise();
   void SaveFinish();
   void RestoreFinish();
-  void SetMC(fixed _mc) { gps_info.MacCready = (double) _mc; }
+  void SetMC(fixed _mc) { gps_info.MacCready = _mc; }
 
   virtual fixed GetAverageThermal() const;
   virtual void OnClimbBase(fixed StartAlt) = 0;
